requirements.txt
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
peft>=0.10.0
trl>=0.8.6
bitsandbytes>=0.43.0
accelerate>=0.28.0
wandb>=0.16.0
openai>=1.0.0        # optional: GPT-4 judge for win-rate eval
scipy
numpy
pyyaml
tqdm
